# Possible performance improvement moving componentStyle.generateAndInjectStyles out of render