--- a/example/devServer.js
+++ b/example/devServer.js
@@ -25,6 +25,10 @@ const port = 3000
 
 app.use(Express.static('dist'))
 
+app.get('/with-perf.html', (req, res) => {
+  res.sendFile(path.join(__dirname, 'with-perf.html'))
+})
+
 app.get('/*', (req, res) => {
   res.sendFile(path.join(__dirname, 'index.html'))
 })