--- a/example/with-perf.html
+++ b/example/with-perf.html
@@ -1,0 +1,84 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>Example with Perf logging</title>
+  </head>
+  <body>
+    <h1>Example with Perf logging</h1>
+    <div id="container"></div>
+    <script src="https://unpkg.com/react@15/dist/react-with-addons.js"></script>
+    <script src="https://unpkg.com/react-dom@15.3.2/dist/react-dom.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/babel-core/5.8.24/browser.min.js"></script>
+    <script src="/styled-components.js"></script>
+    <script type="text/babel">
+
+    const Perf = React.addons.Perf;
+
+    // Create a <Title> react component that renders an <h1> which is
+    // centered, palevioletred and sized at 1.5em
+    const Title = styled.default.h1`
+      font-size: 1.5em;
+      text-align: center;
+      color: palevioletred;
+    `;
+
+    // Create a <Wrapper> react component that renders a <section> with
+    // some padding and a papayawhip background
+    const Wrapper = styled.default.section`
+      padding: 4em;
+      background: papayawhip;
+    `;
+      class Example extends React.Component {
+        render() {
+          return (
+            <Wrapper {...this.props}>
+              <Title>Hello World, this is my first styled component!</Title>
+            </Wrapper>
+          )
+        }
+      }
+
+
+      class ExampleWithPerf extends React.Component {
+        componentWillMount() {
+          Perf.start();
+        }
+
+				stopProfile() {
+					Perf.stop();
+					const measurements = Perf.getLastMeasurements();
+					Perf.printInclusive(measurements);
+				}
+
+        componentDidMount() {
+					this.stopProfile()
+					let i = 0;
+					setInterval(() => {
+						this.setState({
+							i
+						});
+						// i++
+					}, 1000)
+        }
+
+        componentWillReceiveProps() {
+          Perf.start()
+        }
+
+        componentDidUpdate() {
+        	this.stopProfile()
+        }
+
+        render() {
+          return <Example {...this.state} />
+        }
+      }
+
+      ReactDOM.render(
+        <ExampleWithPerf />,
+        document.getElementById('container')
+      );
+    </script>
+  </body>
+</html>