My performance testing experience is quite limited and this is the first time I've used `react-addons-perf`.

While we were perusing the styled-components code we got to wondering about the use of `generateAndInjectStyles` in the `render` method of `StyledComponent`. Our first thought was that a side-effecty method-call probably shouldn't happen in the `render` method but rather in one of the lifecycle hooks such as `componentWillReceiveProps`, therefor keeping `render` a pure method.

I wanted to add some benchmarks to see if the change made any difference as well as trying out a few other things that might have improved performance (playing around with `shouldComponentUpdate` etc).

After putting together a benchmark based on the existing example it looks like there may be a performance improvement made by moving `generateAndInjectStyles` out of `render` and into `componentWillReceiveProps`.

Like I say, my perf experience and knowledge is very limited but I ran these stats over and over to make sure the results were consistent and it seems to be the case. It's also very likely that I'm completely missing something so bare with me on this one.

Here's some numbers from before and after the adjustment. The initial times are the first render. The next two metrics are re-renders after the container hosting the `<Example />` component gets re-rendered due to a state change.

**Before:**
 
<img width="702" alt="with-code-in-render" src="https://cloud.githubusercontent.com/assets/1519443/19661582/2f769e44-9a2c-11e6-8cf8-fc44d5d8eb0e.png">

You can see here that the inclusive render time of `Example > styled.section` (index 2) is `6.54ms` and `Example > styled.h1` (index 3) is `0.68ms`.

**After**

<img width="697" alt="with-code-moved" src="https://cloud.githubusercontent.com/assets/1519443/19661667/7d43514e-9a2c-11e6-9e9b-4e5e42198ed8.png">

After I moved `generateAndInjectStyles` and the other code it's reliant on, out of `render` the inclusive render time of `Example > styled.section` (index 2) has dropped to `0.32ms` and `Example > styled.h1` (index 3) is `0.04ms`.

This seems to suggest a fair performance improvement by shifting the code from `render`.

I've included the benchmark example file that I used to create these results. If I've missed something let me know and feel free to ignore this. If this is of some use then great!