@mxstbr 2016-10-24
Thank you so much for digging into this! I'm not too well-versed in the tradeoffs of this part of the codebase, but @geelen might have something to say here.