@JamieDixon 2016-10-25
I see from the failing tests that I did miss something so I'm having a go at solving that problem whilst retaining the perf numbers outlined.