@JamieDixon 2016-10-25
Ok between myself and @sheepsteak we've got the tests passing and the perf is still looking good. Let me know what feedback you have. I see the travis `danger` task is failing and that might be because we would also need to make changes to the native version also?? Cheers.