@mxstbr 2016-10-28
Released as `styled-components@1.0.10`, twitter announcement coming soon! So many thanks @JamieDixon!